Adding Creature in External Asset. Using new shot cooldown. Adding new base spawn system.

+++ Assets/Script/Enemy/Enemy.cs
@@ -7,7 +7,7 @@ namespace CarAdventure.Entity {
     public class Enemy : MonoBehaviour {
         [Header("Data")]
         [SerializeField]
-        internal int life;
+        internal float life;
         [SerializeField]
         internal float attack;
         [SerializeField]
